Fix: Fixing the erro of ft_putnbr_base that need the base_len to be casted to long int for the comparison to work

+++ src/ft_putnbr_base_fd.c
@@ -27,7 +27,7 @@ size_t	ft_putnbr_base_fd(int n, char *base, int fd)
 		nb *= -1;
 		byte_cont += ft_putchar_fd('-', fd);
 	}
-	if (nb < base_len)
+	if (nb < (long int) base_len)
 		return (ft_putchar_fd(base[nb % base_len], fd));
 	byte_cont += ft_putnbr_base_fd(nb / base_len, base, fd);
 	byte_cont += ft_putnbr_base_fd(nb % base_len, base, fd);
